+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/doc-files/CHANGELOG.md
@@ -16,12 +16,16 @@ exclude: true
 ## v120
 - Added [`disableExtensionProcessSpawning`][120.1] for disabling the extension process spawning. ([bug 1855405]({{bugzilla}}1855405))
 - Added `DisabledFlags.SIGNATURE` for extensions disabled because they aren't correctly signed. ([bug 1847266]({{bugzilla}}1847266))
+- Added `Builder` pattern constructors for [`ReviewAnalysis`][120.2] and [`Recommendation`][120.3] (part of [bug 1846341]({{bugzilla}}1846341))
 
 [120.1]: {{javadoc_uri}}/WebExtensionController.html#disableExtensionProcessSpawning
+[120.2]: {{javadoc_uri}}/GeckoSession.html#ReviewAnalysis.Builder.html
+[120.3]: {{javadoc_uri}}/GeckoSession.html#Recommendation.Builder.html
 
 ## v119
 - Added `remoteType` to GeckoView child crash intent. ([bug 1851518]({{bugzilla}}1851518))
-[119.1]: {{javadoc_uri}}/GeckoSession.html#requestCreateAnalysis(String)
+
+- [119.1]: {{javadoc_uri}}/GeckoSession.html#requestCreateAnalysis(String)
 [119.2]: {{javadoc_uri}}/GeckoSession.html#requestAnalysisCreationStatus(String)
 [119.3]: {{javadoc_uri}}/GeckoSession.html#pollForAnalysisCompleted(String)
 
@@ -1437,4 +1441,4 @@ to allow adding gecko profiler markers.
 [65.24]: {{javadoc_uri}}/CrashReporter.html#sendCrashReport(android.content.Context,android.os.Bundle,java.lang.String)
 [65.25]: {{javadoc_uri}}/GeckoResult.html
 
-[api-version]: 150fd70ec1d59eba6d5354bdbb8f7c54ec7b0dba
+[api-version]: 4e0c739da6386125b4cc45bdb056d8a90f6bf532